remove TaskService interface
